requirements.txt
PyPDF2==3.0.1
et-xmlfile==1.1.0
numpy==1.26.3
openpyxl==3.1.2
pandas==2.1.4
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
tzdata==2023.3
anthropic==0.8.1
python-dotenv==1.0.0
flask==2.3.3
flask-cors==4.0.0
sentence-transformers==2.2.2
selenium==4.11.2
tqdm==4.66.1
openai==1.3.0